# Under ARM, it is slightly better to reverse the word once and then extract the bits.