This gives a nice speedup (about 3%) under ARM to stage 1 over many files.


|   |  before |  after |
|---|---|---|
|  twitter | 5.6936 GB/s  |   5.8474 GB/s |
| canada  | 5.2355 GB/s   |  5.4106 GB/s  |
| apache_builds  |    6.1337 GB/s |  6.3507 GB/s  |
| github_events  |  6.5684 GB/s |   6.7383 GB/s |

These numbers are with clang 12 under an Apple M1 laptop. To get solid results, I use  commands like `benchmark/parse -n 5000 jsonexamples/github_events.json`.

Unfortunately, the required instruction (rbit) is difficult to access from C++. I had to use inline assembly so it does not work under Visual Studio. Visual Studio has such an intrinsic, but it seems limited to 32-bit (at least up to recently).

This is not a large gain but given that we rely heavily on stage 1 for everything, 3% is quite nice.

Fixes https://github.com/simdjson/simdjson/issues/322