@lemire 2021-04-16
Somehow I had closed https://github.com/simdjson/simdjson/issues/322 due to lack of interest and motivation, but I think I was wrong to do so. It is a worthwhile issue.