@aqrit 2021-04-16
possible alternative
https://lemire.me/blog/2019/05/15/bitset-decoding-on-apples-a12/#comment-406914