@lemire 2021-04-17
@aqrit 

Yeah. Ok. So you can do it without bit reversal, but it is not clear to me that it saves you much. Maybe you can help me see more clearly. Here are the functions I have in mind, I simplified them so we could see what is happening.

```C++
#include <cstdint>
// Current code
void flatten_bits(uint32_t *base_ptr, uint32_t *base, uint32_t idx, uint64_t bits) {
  if (bits != 0u) {
    uint32_t cnt = __builtin_popcountll(bits);
    uint32_t next_base = *base + cnt;
    uint32_t b = *base;
    for(uint32_t i = 0; i < 16; i++) {
      base_ptr[b + i] = static_cast<uint32_t>(idx) + __builtin_ctzll(bits);
      bits = bits & (bits - 1);
    }
    *base = next_base;
  }
}

// New PR
void flatten_bits_pr(uint32_t *base_ptr, uint32_t *base, uint32_t idx, uint64_t bits) {
  if (bits != 0u) {
     uint64_t rev_bits;
     __asm("rbit %0, %1" : "=r"(rev_bits) : "r"(bits));
    uint32_t cnt = __builtin_popcountll(bits);
    uint32_t next_base = *base + cnt;
    uint32_t b = *base;    
    for(uint32_t i = 0; i < 16; i++) {
       int lz = __builtin_clzll(rev_bits);
       base_ptr[b + i] = static_cast<uint32_t>(idx) + lz;
       rev_bits = rev_bits ^ (uint64_t(0x8000000000000000) >> lz);
    } 
    *base = next_base;
  }
}

// What some people are proposing???
void flatten_bits_alternative(uint32_t *base_ptr, uint32_t *base, uint32_t idx, uint64_t bits) {
  if (bits != 0u) {
    uint32_t cnt = __builtin_popcountll(bits);
    idx += 63;
    uint32_t next_base = *base + cnt;
    uint32_t b = *base;
    for(uint32_t i = 0; i < 16; i++) {
        uint64_t lowest = (-bits) & (bits );
      base_ptr[b + i] =
       idx - 
       __builtin_clzll(lowest);
      bits -= lowest;
    } 
    *base = next_base;
  }
}



```

Now, here is what godbolt is given me.

https://godbolt.org/z/8E81cdbWj

The current tight loop is:

```asm
.L3:
  rbit x5, x3
  sub x6, x3, #1
  clz x5, x5
  and x3, x3, x6
  add w5, w2, w5
  str w5, [x0, w4, uxtw 2]
  add w4, w4, 1
  cmp w7, w4
  bne .L3
```

Notice how it keep reversing the bits and then calling the leading zero? That's the weird way of computing the trailing zeros. Uh!

I am going to omit the cmp, the bne (jump), the store and the counter increment (add w4, w4, 1).  The store I cannot avoid, the pointer increment can go away under unrolling.

So I have rbit, sub, clz, and, add... so 5 instructions.

Let us compare with x64:

```asm
.L3:
  xor edx, edx
  add rax, 4
  tzcnt rdx, rcx
  blsr rcx, rcx
  add edx, esi
  mov DWORD PTR [rax-4], edx
  cmp rax, rdi
  jne .L3
```

I think `xor edx, edx`  is a noop.  If I ignore the increment, the store and the comparaison, I have tzcnt, blsr, add so only three instructions!  Quite a bit better than ARM.


Ok. Can ARM catch up?... let me look at code matching my PR...

```asm
.L11:
  clz x0, x3
  add w5, w2, w0
  str w5, [x6, w4, uxtw 2]
  add w4, w4, 1
  lsr x0, x8, x0
  eor x3, x3, x0
  cmp w7, w4
  bne .L11
```

So we are introducing a shift which is not so nice, but omitting the cmp and the branch and the load/store, I have 4instructions, which is down from 5 before. So I saved one instruction in the tight loop. Victory!

To be clear, I am paying for the this by the bit reversal that I do once at the beginning of the function... but that's only one instruction. I actually have a relatively fat function header with 10 instructions or so... hence it is not going to make a bit difference. I also have a fast path when the input word is zero.

The essential instructions are...

```
  clz x0, x3 // compute __builtin_clzll(rev_bits);
  add w5, w2, w0 // compute the sum static_cast<uint32_t>(idx) + lz;
 lsr x0, x8, x0 // compute (uint64_t(0x8000000000000000) >> lz)
  eor x3, x3, x0 // compute the xor in rev_bits ^ (uint64_t(0x8000000000000000) >> lz)
```

Can we go lower?


What about what I think you are proposing?

```asm
.L19:
  neg x2, x3
  and x2, x2, x3
  clz x5, x2
  sub x3, x3, x2
  sub w2, w6, w5
  str w2, [x0, w4, uxtw 2]
  add w4, w4, 1
  cmp w7, w4
  bne .L19
```

I go back to 5 instructions.

That's instruction count. What about data parallelism?


The x64 updates the word with one instruction (  blsr rcx, rcx ), so it can pipeline a ton of instructions per cycle. In the main branch, ARM has two critical instructions per cycle...

```
  sub x6, x3, #1
  and x3, x3, x6
```

This is what keeps x3 updated. So unless there is fusion, it needs two cycles at least per loop.

The new code in this PR is much worse... it has a sequence of three instructions that are data dependent... that's not good.

```
  clz x0, x3
  lsr x0, x8, x0
  eor x3, x3, x0
```

So it is not obvious that the PR would do better... but it does empirically.

But what I think you are proposing is just as bad...
```
  neg x2, x3
  and x2, x2, x3
  sub x3, x3, x2
```