@lemire 2021-04-17
This is pretty:

```C++
void flatten_bits_alternative(uint32_t *base_ptr, uint32_t *base, uint32_t idx, uint64_t bits) {
  if (bits != 0u) {
    uint32_t cnt = __builtin_popcountll(bits);
    idx += 63;
    uint32_t next_base = *base + cnt;
    uint32_t b = *base;
    uint64_t lowest = 0;
    for(uint64_t i = 0; i < 16; i++) {
        lowest = (bits - lowest)
         & (lowest - bits);
      base_ptr[b + i] =
       idx - __builtin_clzll(lowest);
    } 
    *base = next_base;
  }
}
```

But it is still generate 5 non-trivial instruction with a sequence of 3 data dependent instruction. So it is not obviously better than the PR with is 4 non-trivial instructions.