@aqrit 2021-04-17
CLZ and variable shift are fast on ARM.
So figure the reversed version has a 3 cycle latency chain
```
for(...){
    lz = CLZ(bits); // cycle 1
    hi_bit = 0x8000000000000000 >> lz; // cycle 2
    bits ^= hi_bit; // cycle 3

    index_buf[i] = offset + lz; // off chain
}
```

the alternative version is one more instruction in the inner loop, but only has 2 cycles on the critical path (maybe?)
```
static inline
void flatten_bits_alternative(uint64_t bits, uint32_t offset, uint32_t* index_buf)
{
	uint64_t next_bit, temp;
	uint32_t popcnt; 

	popcnt = __builtin_popcountll(bits);
	next_bit = bits & (-bits);
	index_buf[0] = offset - CLZ(next_bit);

	for(int i = 1; i < 8; i++) {
		temp = next_bit - bits; // cycle 1
		bits = bits ^ next_bit; // cycle 1
		next_bit = temp & bits; // cycle 2

		index_buf[i] = offset - CLZ(next_bit); // off chain?
	}

	if (popcnt > 8) {
		for(int i = 0; i < 8; i++) {
			; //...
		}
	}
	//... 
}
```