@aqrit 2021-04-17
With the reversed version, I wonder if it would be better to just shift off the highest set bit directly:
```
for(...){
    lz = CLZ(bits);
    bits <<= lz;
    offset += lz;
    index_buf[i] = offset; 
}
```