@lemire 2021-04-30
This PR can probably be merged.